% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{characterORNULL-class}
\alias{characterORNULL-class}
\title{Class Union: character or NULL}
\description{
Internal union class used in slots that may contain
a character vector or be \code{NULL}.
}
\keyword{internal}
