% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.r
\name{check_beta_gamma}
\alias{check_beta_gamma}
\title{Check consistency of beta and gamma matrices}
\usage{
check_beta_gamma(beta, gamma, q, case = "ns")
}
\arguments{
\item{beta}{Beta coefficient matrix.}

\item{gamma}{Covariance matrix of random effects.}

\item{q}{Number of parameters (subject or not subject to selection).}

\item{case}{Character; either "ns" (not to select) or "s" (to select).}
}
\description{
Check consistency of beta and gamma matrices
}
\keyword{internal}
