% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.r
\name{check_beta_hdim}
\alias{check_beta_hdim}
\title{Check beta matrix for high-dimensional parameters}
\usage{
check_beta_hdim(beta, pv)
}
\arguments{
\item{beta}{Beta coefficient matrix.}

\item{pv}{Number of parameters to select (including intercept).}
}
\description{
Check beta matrix for high-dimensional parameters
}
\keyword{internal}
