% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.r
\name{check_beta_support}
\alias{check_beta_support}
\title{Check beta matrix dimensions against support}
\usage{
check_beta_support(beta, support)
}
\arguments{
\item{beta}{Beta coefficient matrix.}

\item{support}{Covariate support matrix (can be NULL).}
}
\description{
Check beta matrix dimensions against support
}
\keyword{internal}
