% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-check.r
\name{check_data}
\alias{check_data}
\title{Check Data Consistency with Model}
\usage{
check_data(data, model)
}
\arguments{
\item{data}{An object of class \code{saemvsData}.}

\item{model}{An object of class \code{saemvsModel}.}
}
\description{
Internal function to verify that the data provided by the user
is consistent with the SAEMVS model specification.
}
\details{
Performs the following checks:
\itemize{
\item If \code{phi_to_select_idx} is non-empty, \code{x_candidates} must be provided.
\item \code{x_candidates}, if present, must have one row per sequence in \code{y_series}.
\item \code{x_forced} and \code{x_forced_support}, if present, must be numeric matrices
with compatible dimensions.
}
}
\keyword{internal}
