% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-check.r
\name{check_hyper}
\alias{check_hyper}
\title{Check Hyperparameters Consistency}
\usage{
check_hyper(hyper, model, tuning)
}
\arguments{
\item{hyper}{An object of class \code{saemvsHyperSpikeAndSlab}.}

\item{model}{An object of class \code{saemvsModel}.}

\item{tuning}{An object of class \code{saemvsTuning}.}
}
\description{
Internal function to verify that the hyperparameters for the spike-and-slab
prior are consistent with the SAEMVS model and tuning parameters.
}
\details{
Performs the following checks:
\itemize{
\item \code{inclusion_prob_prior_a} and \code{inclusion_prob_prior_b}
lengths must match the number of phi parameters subject to selection.
\item \code{cov_re_prior_scale} must be a numeric square matrix with
dimensions equal to the number of phi parameters selected.
\item All values in \code{spike_values_grid} must be smaller than
\code{slab_parameter}.
}
}
\keyword{internal}
