% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-check.r
\name{check_init}
\alias{check_init}
\title{Check Initial Values Consistency}
\usage{
check_init(init, data, model)
}
\arguments{
\item{init}{An object of class \code{saemvsInit}.}

\item{data}{An object of class \code{saemvsData}.}

\item{model}{An object of class \code{saemvsProcessedModel}.}
}
\description{
Internal function to verify that the initial values provided by the user
match the SAEMVS model and data dimensions.
}
\details{
Performs the following checks:
\itemize{
\item Length of \code{intercept} must match \code{phi_dim} in the model.
\item \code{beta_candidates} and \code{beta_forced}, if present, must have
the same number of columns as \code{intercept}.
\item \code{cov_re} must be a numeric square matrix of size \code{phi_dim}.
\item \code{sigma2} must be a single strictly positive number.
}
}
\keyword{internal}
