% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.r
\name{check_positive_or_null_slot}
\alias{check_positive_or_null_slot}
\title{Check that a numeric slot is either NULL or strictly positive}
\usage{
check_positive_or_null_slot(slot, name_slot)
}
\arguments{
\item{slot}{Numeric value or NULL.}

\item{name_slot}{Name of the slot.}
}
\description{
Check that a numeric slot is either NULL or strictly positive
}
\keyword{internal}
