% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.r
\name{check_positive_slot}
\alias{check_positive_slot}
\title{Check that a numeric slot is strictly positive}
\usage{
check_positive_slot(slot, name_slot)
}
\arguments{
\item{slot}{Numeric value.}

\item{name_slot}{Name of the slot.}
}
\description{
Check that a numeric slot is strictly positive
}
\keyword{internal}
