% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{df_cov}
\alias{df_cov}
\title{Example dataset: df_cov}
\format{
A data frame with one row per individual and 200 covariate columns.
}
\usage{
data(df_cov)
}
\description{
Simulated covariate dataset for \code{saemvs}.
}
\details{
Contains 200 candidate covariates in wide format, with one row per
individual.
}
