% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-predict.r
\name{.predict_one_support}
\alias{.predict_one_support}
\title{Compute individual-level predictions for one support}
\usage{
.predict_one_support(object, support_idx, k = 20)
}
\arguments{
\item{object}{A \code{saemvsResults} object.}

\item{support_idx}{Integer. Index of the support to use.}

\item{k}{Integer. Number of last SAEM iterations to average.}
}
\value{
A numeric matrix with:
\itemize{
\item rows corresponding to individuals
\item columns corresponding to \eqn{\phi} parameters
}
}
\description{
Internal utility function used by \code{predict()} and
\code{predict_support()}. For a given support, it averages the last
\eqn{k} SAEM iterations of the individual parameters \eqn{\phi}.
}
\details{
Let \eqn{\phi^{(t)}_{i,p}} denote the value of parameter \eqn{p} for
individual \eqn{i} at SAEM iteration \eqn{t}. This function returns:
\deqn{
\hat{\phi}_{i,p} = \frac{1}{k} \sum_{t=T-k+1}^{T} \phi^{(t)}_{i,p},
}
where \eqn{T} is the total number of iterations for the selected support.
}
\keyword{internal}
