% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{example_data_df}
\alias{example_data_df}
\title{Example datasets for saemvs}
\usage{
data(example_data_df)
}
\description{
Simulated datasets used to demonstrate the usage of the \code{saemvs} package
for variable selection in nonlinear mixed-effects models.
}
\details{
The \code{.rda} file contains two data frames:
\itemize{
\item \code{df_long}: longitudinal measurements in long format
\item \code{df_cov}: covariates in wide format
}
}
