% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{listORNULL-class}
\alias{listORNULL-class}
\title{Class Union: list or NULL}
\description{
Internal union class used in slots that may contain
a list or be \code{NULL}.
}
\keyword{internal}
