% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-config.r
\name{make_config}
\alias{make_config}
\title{Create a unified configuration list for the SAEMVS algorithm}
\usage{
make_config(data, model, tuning_algo, init, hyperparam)
}
\arguments{
\item{data}{A \code{saemvsProcessedData} object.}

\item{model}{A \code{saemvsModel} object.}

\item{tuning_algo}{A \code{saemvsTuning} object.}

\item{init}{A \code{saemvsProcessedInit} object.}

\item{hyperparam}{A \code{saemvsHyperSlab} object.}
}
\value{
A named list containing all information necessary to run the SAEMVS algorithm.
}
\description{
This function constructs a configuration list containing processed data, model information,
initialization, hyperparameters, and tuning parameters, with fields aligned with the
original object slots.
}
\keyword{internal}
