% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{matrixORNULL-class}
\alias{matrixORNULL-class}
\title{Class Union: matrix or NULL}
\description{
Internal union class used in slots that may contain
a numeric matrix or be \code{NULL}.
}
\keyword{internal}
