% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{numericORNULL-class}
\alias{numericORNULL-class}
\title{Class Union: numeric or NULL}
\description{
Internal union class used in slots that may contain
a numeric vector or be \code{NULL}.
}
\keyword{internal}
