% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print.r
\name{print,saemvsResults-method}
\alias{print,saemvsResults-method}
\title{Print a Summary of SAEM Variable Selection Results}
\usage{
\S4method{print}{saemvsResults}(x)
}
\arguments{
\item{x}{An object of class \code{saemvsResults} containing the output of
a SAEM variable selection run.}
}
\description{
Displays a concise overview of the results from a SAEM variable selection
procedure. The output provides high-level information about the model
selection process and identifies the best model according to the chosen
selection criterion.
}
\details{
The printed summary includes:
\itemize{
\item The selection criterion used (BIC or e-BIC).
\item The number of spike parameter values explored.
\item The number of candidate supports evaluated.
\item The number of unique supports retained after removing duplicates.
\item The criterion value of the best selected model.
\item The number of covariates selected in the best model.
}

For detailed parameter estimates and selected variables, use
\code{\link{summary}}.
}
\examples{
\dontrun{
res <- saemvs(...)
print(res)
}

}
