% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{saemResults-class}
\alias{saemResults-class}
\title{Classe saemResults}
\description{
Internal class storing intermediate estimation results produced by
the SAEM algorithm. Objects of this class are not meant
to be created by the user but are automatically produced by
\code{saemvs}.
}
\section{Slots}{

\describe{
\item{\code{beta_to_select}}{A list of matrices containing the estimated regression
coefficients for covariates that are related with the components of
\code{phi} that are subject to selection (\code{phi_to_select_idx}) at each
iteration.}

\item{\code{beta_not_to_select}}{A list of matrices containing the estimated
regression coefficients for covariates that are related with the components
of \code{phi} that are not subject to selection (\code{phi_not_to_select_idx}) at
each iteration.}

\item{\code{gamma_to_select}}{A list of covariance matrices estimated at each
iteration for the random effects associated with \code{phi_to_select}.}

\item{\code{gamma_not_to_select}}{A list of covariance matrices estimated at each
iteration for the random effects associated with \code{phi_not_to_select}.}

\item{\code{sigma2}}{A numeric vector of estimated residual variances at each
iteration.}

\item{\code{phi_to_select_idx}}{Numeric vector or NULL. Indices of components of
\eqn{\phi} that are subject to variable selection.}

\item{\code{phi_not_to_select_idx}}{Numeric vector or NULL. Indices of components of
\eqn{\phi} that are not subject to variable selection.}

\item{\code{phi_names}}{Character vector or NULL. Names of all \eqn{\phi} components,
used for labeling and summaries.}

\item{\code{x_candidates_names}}{Character vector or NULL. Names of candidate
covariates associated with \code{beta_to_select}.}

\item{\code{x_forced_names}}{Character vector or NULL. Names of forced covariates
associated with \code{beta_not_to_select}.}
}}

