% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\name{saemvsData}
\alias{saemvsData}
\title{Constructor for saemvsData}
\usage{
saemvsData(y, t, x_candidates = NULL, x_forced = NULL)
}
\arguments{
\item{y}{list of numeric vectors (responses).}

\item{t}{list of numeric vectors (time indices).}

\item{x_candidates}{matrix or NULL, candidate covariates for selection.}

\item{x_forced}{matrix or NULL, forced covariates.}
}
\value{
An object of class \code{saemvsData}.
}
\description{
Constructor for saemvsData
}
\keyword{internal}
