% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{saemvsHyperSpikeAndSlab-class}
\alias{saemvsHyperSpikeAndSlab-class}
\title{Class "saemvsHyperSpikeAndSlab"}
\description{
An S4 class that extends \code{\linkS4class{saemvsHyperSlab}} by adding
a spike hyperparameter for the spike-and-slab prior.
}
\section{Slots}{

\describe{
\item{\code{spike_parameter}}{\code{numeric}. The spike parameter, must be strictly
positive.}
}}

\section{Prototype}{

Defaults to \code{spike_parameter = 0.1}.
}

\section{Validity}{

Ensures that \code{spike_parameter} is strictly positive.
}

\seealso{
\code{\linkS4class{saemvsHyperSlab}}
}
\keyword{internal}
