% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\name{saemvsHyperSpikeAndSlab}
\alias{saemvsHyperSpikeAndSlab}
\title{Internal constructor for saemvsHyperSpikeAndSlab}
\usage{
saemvsHyperSpikeAndSlab(spike_parameter, hyper_slab)
}
\arguments{
\item{spike_parameter}{numeric, strictly positive}

\item{hyper_slab}{an object of class \code{saemvsHyperSlab}}
}
\value{
An object of class \code{saemvsHyperSpikeAndSlab}
}
\description{
Internal constructor for saemvsHyperSpikeAndSlab
}
\keyword{internal}
