% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\name{saemvsProcessedData}
\alias{saemvsProcessedData}
\title{Internal constructor for saemvsProcessedData}
\usage{
saemvsProcessedData(
  x_phi_to_select = NULL,
  x_phi_not_to_select = NULL,
  tx_x_phi_to_select = NULL,
  kron_tx_x_phi_to_select = NULL,
  x_phi_not_to_select_list = NULL,
  ...
)
}
\description{
Internal constructor for saemvsProcessedData
}
\keyword{internal}
