% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{saemvsProcessedModel-class}
\alias{saemvsProcessedModel-class}
\title{Internal processed SAEMVS model}
\description{
Internal representation of a SAEMVS model used by the estimation and
variable selection algorithms.

This class is not intended to be created or manipulated directly by users.
It stores a fully processed version of a \code{saemvsModel}, where all
parameters are indexed numerically and forced covariate information is
encoded as a binary design matrix.
}
\section{Slots}{

\describe{
\item{\code{model_func}}{A function of the form \code{function(phi, t)} returning
predicted values for given parameter vector \code{phi} and time \code{t}.}

\item{\code{phi_dim}}{Integer. Total number of model parameters (\code{phi}).}

\item{\code{phi_to_select_idx}}{Integer vector or \code{NULL}. Indices of parameters
on which covariate selection is performed.}

\item{\code{phi_fixed_idx}}{Integer vector or \code{NULL}. Indices of parameters that
are fixed (i.e. without random variability).}

\item{\code{x_forced_support}}{Numeric matrix or \code{NULL}. Binary support matrix
encoding forced covariate effects. Columns correspond to model parameters
(length \code{phi_dim}); rows correspond to covariates. A value of 1
indicates that the covariate is forced for the corresponding parameter.}
}}

\keyword{internal}
