% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{small_example_data}
\alias{small_example_data}
\title{Small example dataset for saemvs}
\format{
A data frame with 100 individuals and 4 covariates:
\describe{
\item{y_list}{Responses. List of numeric vectors, each one containing observed responses for one individual.}
\item{t_list}{Time points. List of numeric vectors, each one containing observation time points for one individual.}
\item{x}{Matrix of covariates with dimensions 100 x 4 (rows = individuals, columns = covariates).}
}
}
\usage{
data(small_example_data)
}
\description{
A small simulated dataset to demonstrate the usage of \code{saemvs}.
This dataset is intended to illustrate the usage of the \code{saemvs} package for variable selection in nonlinear mixed-effects models.
}
