% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summary.r
\name{summary_all}
\alias{summary_all}
\title{Summary of all supports}
\usage{
summary_all(object, digits = 3)
}
\arguments{
\item{object}{A \code{saemvsResults} object}

\item{digits}{Number of digits for printing numeric values (default = 3)}
}
\description{
Iterates through all unique supports in a SAEMVS run and prints a
summary for each.
}
