% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summary.r
\name{summary_support}
\alias{summary_support}
\title{Summary of a specific support}
\usage{
summary_support(object, support_idx, digits = 3)
}
\arguments{
\item{object}{A \code{saemvsResults} object}

\item{support_idx}{Integer. Index of the support to display in
\code{unique_support}.}

\item{digits}{Integer (default = 3). Number of decimal places
used when printing coefficients and the covariance matrix.}
}
\description{
Provides a summary of the model corresponding to a given support
obtained from a SAEM variable
selection run.
Displays:
\itemize{
\item Selected covariates for each \eqn{\varphi} parameter
\item Estimated regression coefficients (\eqn{\beta})
\item Estimated covariance matrix (\eqn{\Gamma})
}
}
