% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-saemvs.r
\name{test_saemvs}
\alias{test_saemvs}
\title{Run a Single SAEMVS Test Fit}
\usage{
test_saemvs(data, model, init, tuning_algo, hyperparam)
}
\arguments{
\item{data}{A \link[=saemvsData-class]{saemvsData} object containing the observed response series.}

\item{model}{A \link[=saemvsModel-class]{saemvsModel} object defining the model structure.}

\item{init}{A \link[=saemvsInit-class]{saemvsInit} object providing initial parameter values.}

\item{tuning_algo}{A \link[=saemvsTuning-class]{saemvsTuning} object specifying algorithm hyperparameters.}

\item{hyperparam}{A \link[=saemvsHyperSlab-class]{saemvsHyperSlab} object defining the slab prior.}
}
\value{
A \link[=saemResults-class]{saemResults} object containing estimated parameters.
}
\description{
Executes a simplified SAEMVS procedure for testing purposes.
}
\details{
This function allows a quick fit of SAEMVS to inspect the evolution of parameter estimates.
The function performs the following steps:
\enumerate{
\item Compile the R model function to C++ using \code{compile_model}.
\item Construct a spike-and-slab hyperparameter object.
\item Run the SAEM algorithm using \code{run_saem}.
\item Return parameter estimates packaged in a \code{saemResults} object.
}
}
\examples{
\dontrun{
test_results <- test_saemvs(data_obj, model_obj, init_obj, tuning_obj, hyper_obj)
summary(test_results)
}

}
