% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{df_long}
\alias{df_long}
\title{Example dataset: df_long}
\format{
A data frame with columns \code{id}, \code{time}, and \code{y}.
}
\usage{
data(df_long)
}
\description{
Long-format simulated longitudinal data for \code{saemvs}.
}
\details{
Each row corresponds to one observation for one individual at a given
time point.
}
