% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m-steps.r
\name{m_step_mle}
\alias{m_step_mle}
\title{Internal M-step: update MLE parameters}
\usage{
m_step_mle(config, k, state)
}
\arguments{
\item{config}{List containing model configuration and hyperparameters.}

\item{k}{Current iteration index (integer).}

\item{state}{List containing the current state of the algorithm.}
}
\value{
Updated \code{state} list with new sigma2, beta and gamma.
}
\description{
Updates the residual variance (sigma2) and parameters in the MLE case.
}
\keyword{internal}
