% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{saemvsModel-class}
\alias{saemvsModel-class}
\title{Class saemvsModel}
\description{
Represents a model for SAEMVS, including the model function and
indexing information for variable selection on parameters phi.
}
\section{Slots}{

\describe{
\item{\code{model_func}}{A function of form function(phi, t) returning predicted values.}

\item{\code{phi_dim}}{Integer: total number of phi parameters.}

\item{\code{phi_to_select_idx}}{Integer vector or NULL: indices of phi parameters
on which variable selection will be performed.}

\item{\code{phi_fixed_idx}}{Integer vector or NULL: indices of phi parameters
that are fixed (i.e. without random variability).}

\item{\code{x_forced_support}}{Numeric matrix or NULL: design matrix for covariates
that are forced in the model. Must have phi_dim columns.}
}}

